requirements.txt
git+https://github.com/mongodb/sonar@0.0.11#egg=sonar
docker==4.3.1
kubernetes==11.0.0
jinja2==2.11.3
PyYAML==5.4.1
black==20.8b1
mypy==0.782
tqdm==v4.49.0
boto3==1.16.21
pymongo==3.11.4
dnspython==2.0.0
requests==2.24.0
pyyaml==5.4.1
rsa>=4.7 # not directly required, pinned by Snyk to avoid a vulnerability
